let RestartingUsrp inherit from C++ USRP
by doing that, the aggregation of pybinding.Usrp into RestartingUsrp is not needed
and hence no forwarding-calls need to be written. Therefore, one layer of call forwarding
is omitted.

The three-times testing if the USRP can be created is obtained by adding a static
factory function to the RestartingUsrp class which tries to instantiate the Usrp 3 times
before it gives up.